+++ backend/api/src/main/kotlin/io/tolgee/hateoas/quickStart/QuickStartModelAssembler.kt
@@ -1,15 +1,15 @@
 package io.tolgee.hateoas.quickStart
 
-import io.tolgee.model.QuickStart
+import io.tolgee.dtos.queryResults.organization.IQuickStart
 import org.springframework.hateoas.server.mvc.RepresentationModelAssemblerSupport
 import org.springframework.stereotype.Component
 
 @Component
-class QuickStartModelAssembler() : RepresentationModelAssemblerSupport<QuickStart, QuickStartModel>(
-  QuickStart::class.java,
+class QuickStartModelAssembler : RepresentationModelAssemblerSupport<IQuickStart, QuickStartModel>(
+  IQuickStart::class.java,
   QuickStartModel::class.java,
 ) {
-  override fun toModel(entity: QuickStart): QuickStartModel {
+  override fun toModel(entity: IQuickStart): QuickStartModel {
     return QuickStartModel(
       finished = entity.finished,
       completedSteps = entity.completedSteps.toMutableList(),