+++ backend/data/src/main/kotlin/io/tolgee/dtos/queryResults/organization/IQuickStart.kt
@@ -1,0 +1,7 @@
+package io.tolgee.dtos.queryResults.organization
+
+interface IQuickStart {
+  val finished: Boolean
+  val completedSteps: Array<String>
+  val open: Boolean
+}