+++ backend/data/src/main/kotlin/io/tolgee/dtos/queryResults/organization/QuickStartView.kt
@@ -1,0 +1,7 @@
+package io.tolgee.dtos.queryResults.organization
+
+class QuickStartView(
+  override val finished: Boolean,
+  override val completedSteps: Array<String>,
+  override val open: Boolean,
+) : IQuickStart