+++ backend/data/src/main/kotlin/io/tolgee/repository/OrganizationRepository.kt
@@ -1,9 +1,9 @@
 package io.tolgee.repository
 
+import io.tolgee.dtos.queryResults.organization.OrganizationView
 import io.tolgee.model.Organization
 import io.tolgee.model.UserAccount
 import io.tolgee.model.enums.OrganizationRoleType
-import io.tolgee.model.views.OrganizationView
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.data.jpa.repository.JpaRepository
@@ -22,9 +22,18 @@ interface OrganizationRepository : JpaRepository<Organization, Long> {
   fun findBySlug(slug: String): Organization?
 
   @Query(
-    """select distinct o as organization, r.type as currentUserRole
+    """select distinct new io.tolgee.dtos.queryResults.organization.OrganizationView(
+          o.id,
+          o.name,
+          o.slug,
+          o.description,
+          bp._scopes,
+          bp.type,
+          r.type,
+          o.avatarHash
+        )
         from Organization o 
-        join fetch o.basePermission as bp
+        join o.basePermission bp
         left join OrganizationRole r on r.user.id = :userId
           and r.organization = o and (r.type = :roleType or :roleType is null)
         left join o.projects p on p.deletedAt is null
@@ -115,9 +124,18 @@ interface OrganizationRepository : JpaRepository<Organization, Long> {
   fun findUsersDefaultOrganization(user: UserAccount): Organization?
 
   @Query(
-    """select distinct o as organization, r.type as currentUserRole
+    """select distinct new io.tolgee.dtos.queryResults.organization.OrganizationView(
+          o.id,
+          o.name,
+          o.slug,
+          o.description,
+          bp._scopes,
+          bp.type,
+          r.type,
+          o.avatarHash
+        )
         from Organization o
-        join fetch o.basePermission bp
+        join o.basePermission bp
         left join OrganizationRole r on r.user.id = :userId and r.organization = o
         where (:search is null or (lower(o.name) like lower(concat('%', cast(:search as text), '%'))))
         and o.deletedAt is null
@@ -177,4 +195,27 @@ interface OrganizationRepository : JpaRepository<Organization, Long> {
   """,
   )
   fun find(id: Long): Organization?
+
+  @Query(
+    """
+    select new io.tolgee.dtos.queryResults.organization.OrganizationView(
+      o.id,
+      o.name,
+      o.slug,
+      o.description,
+      bp._scopes,
+      bp.type,
+      r.type,
+      o.avatarHash
+    )
+    from Organization o
+    join o.basePermission bp
+    left join o.memberRoles r on r.user.id = :currentUserId 
+    where o.id = :id and o.deletedAt is null
+  """,
+  )
+  fun findView(
+    id: Long,
+    currentUserId: Long,
+  ): OrganizationView?
 }