+++ backend/data/src/main/kotlin/io/tolgee/repository/QuickStartRepository.kt
@@ -1,7 +1,9 @@
 package io.tolgee.repository
 
+import io.tolgee.dtos.queryResults.organization.QuickStartView
 import io.tolgee.model.QuickStart
 import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
 
 @Repository
@@ -12,4 +14,19 @@ interface QuickStartRepository : JpaRepository<QuickStart, Long> {
     userAccountId: Long,
     organizationId: Long,
   ): QuickStart?
+
+  @Query(
+    """
+    select new io.tolgee.dtos.queryResults.organization.QuickStartView(
+      qs.finished,
+      qs.completedSteps,
+      qs.open
+    ) from QuickStart qs
+    where qs.userAccount.id = :userAccountId and qs.organization.id = :organizationId
+    """,
+  )
+  fun findView(
+    userAccountId: Long,
+    organizationId: Long,
+  ): QuickStartView?
 }