+++ e2e/cypress/e2e/organizations/organizationsList.cy.ts
@@ -7,10 +7,11 @@ import {
   switchToOrganization,
 } from '../../common/shared';
 import { organizationTestData } from '../../common/apiCalls/testData/testData';
-import { login } from '../../common/apiCalls/common';
+import { login, setBypassSeatCountCheck } from '../../common/apiCalls/common';
 
 describe('Organization List', () => {
   beforeEach(() => {
+    setBypassSeatCountCheck(true);
     organizationTestData.clean();
     organizationTestData
       .generate()
@@ -20,6 +21,10 @@ describe('Organization List', () => {
 
   afterEach(() => {
     organizationTestData.clean();
+  });
+
+  afterEach(() => {
+    setBypassSeatCountCheck(false);
   });
 
   it('creates organization', () => {