+++ e2e/cypress/e2e/projects/projectsListDashboard.cy.ts
@@ -2,10 +2,11 @@ import { HOST } from '../../common/constants';
 import 'cypress-file-upload';
 import { assertTooltip, gcy } from '../../common/shared';
 import { projectListData } from '../../common/apiCalls/testData/testData';
-import { login } from '../../common/apiCalls/common';
+import { login, setBypassSeatCountCheck } from '../../common/apiCalls/common';
 
 describe('Projects Dashboard', () => {
   beforeEach(() => {
+    setBypassSeatCountCheck(true);
     projectListData.clean();
     projectListData.generate();
     login('projectListDashboardUser', 'admin');
@@ -47,6 +48,7 @@ describe('Projects Dashboard', () => {
 
   afterEach(() => {
     projectListData.clean();
+    setBypassSeatCountCheck(false);
   });
 });